+++ twote-web/src/App.css
@@ -1,9 +1,6 @@
 .Middle-Section {
-  padding: 20px;
   display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  align-items: center;
+  flex-direction: column;
   width: 100%;
 }
